refac: extracted blockservice out of learningunitservice
