+++ backend/src/main/java/ch/nova_omnia/lernello/api/LearningUnitRestController.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
+import ch.nova_omnia.lernello.service.BlockService;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
@@ -38,6 +39,7 @@ public class LearningUnitRestController {
     private final LearningUnitService learningUnitService;
     private final LearningUnitMapper learningUnitMapper;
     private final TemporaryKeyMapper temporaryKeyMapper;
+    private final BlockService blockService;
 
     @GetMapping("/{id}")
     @PreAuthorize("hasAuthority('SCOPE_learningUnit:read')")
@@ -52,11 +54,10 @@ public class LearningUnitRestController {
         return id;
     }
 
-
     @PostMapping("/{id}/apply-block-actions")
     @PreAuthorize("hasAuthority('SCOPE_learningUnit:write')")
     public @Valid Map<String, UUID> applyBlockActions(@PathVariable UUID id, @RequestBody List<BlockActionDTO> actionQueue) {
-        Map<String, UUID> temporaryKeyMap = learningUnitService.applyBlockActions(id, actionQueue);
+        Map<String, UUID> temporaryKeyMap = blockService.applyBlockActions(id, actionQueue);
         return temporaryKeyMapper.toDTO(temporaryKeyMap).temporaryKeyMap();
     }
